% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/05_ged_table.R
\name{ged_table}
\alias{ged_table}
\title{ged_table()}
\usage{
ged_table(enroll_data, pos_data, measure = c("enrolled in GED",
  "assessed on academic skills", "gained academic skills", "took GED",
  "passed GED"), target = c(enrolled = 0.75, assessed = 0.9, gained = 0.7,
  took = 0.5, passed = 0.8), eto_programs = c("ss - ged",
  "pg - employment ged", "dc - wise ged"),
  workforce_programs = c("ss - ccorps projects", "ss - counseling",
  "ss - ged", "ss - job placement", "ss - job readiness",
  "pg - employment case management", "pg - employment ged",
  "pg - employment in school", "pg - employment job placement",
  "pg - employment job training", "dc - wise ged", "dc - wise job placement",
  "dc - wise job training"))
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data."[Admin] raw_enrollment_report"}

\item{pos_data}{dataframe: a dataframe containing point of service data."[Admin] raw_pos_report_detailed"}

\item{measure}{character vector: Name of measures in the table}

\item{target}{numeric vector: Organization target for each measure in the table}

\item{eto_programs}{character vector:  a vector of character containing the name of ETO programs to keep for GED analysis.}

\item{workforce_programs}{character vector:  a vector of character containing the name of all workforce programs.}
}
\value{
dataframe
}
\description{
This function creates a summary table of GED results.
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
enroll <- laycUtils::format_data(enroll)
tp <- laycUtils::load_txt('./my_data_folder/touchpoints.txt')
tp <- laycUtils::format_data(tp)

ged_table(enroll_data = enroll, pos_data = tp)
}

