% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{id_prepost}
\alias{id_prepost}
\title{id_prepost()}
\usage{
id_prepost(df)
}
\arguments{
\item{df}{dataframe: a dataframe containing longitudinal data.}
}
\value{
dataframe
}
\description{
This is a helper function that identifies 'pre' and 'post' test based on the date the test was taken.
}
\examples{
tp <- laycUtils::load_txt('./my_data_folder/touchpoints.txt')
tp <- laycUtils::format_data(tp)
tp <- tp[tp$tp_name == "jrt pre-/post test", ]

id_prepost(df = tp)
}

