% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/02_job_placed.R
\name{job_placed}
\alias{job_placed}
\title{job_placed()}
\usage{
job_placed(enroll_data, job_data, eto_programs = c("ss - job placement",
  "pg - employment job placement", "dc - wise job placement",
  "dc - wise job training", "pg - employment job training",
  "pg - employment in school", "pg - employment case management",
  "pg - employment job placement", "ss - job readiness",
  "ss - ccorps projects"))
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data."[Admin] raw_enrollment_report"}

\item{job_data}{dataframe: a dataframe containing employment data."[Admin] raw_job_report"}

\item{eto_programs}{character vector:  a vector of character containing the name of ETO programs to keep for analysis.}
}
\value{
numeric
}
\description{
This function returns the number of participants who enrolled in job placement. It takes one dataframe as input. The data comes from the following ETO Results report: "[Admin] raw_enrollment_report".
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
enroll <- laycUtils::format_data(enroll)
job <- laycUtils::load_txt('./my_data_folder/job.txt')
job <- laycUtils::format_data(job)

job_placed(enroll_data = enroll)
}

