% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/job_table.R
\name{job_table}
\alias{job_table}
\title{job_table()}
\usage{
job_table(enroll_data, job_data, measure = c("enrolled in placement",
  "placed", "retained"), target = c(enrolled = 0.75, placed = 0.4, retained =
  0.7), eto_programs = c("ss - job placement",
  "pg - employment job placement", "dc - wise job placement"),
  workforce_programs = c("ss - ccorps projects", "ss - counseling",
  "ss - ged", "ss - job placement", "ss - job readiness",
  "pg - employment case management", "pg - employment ged",
  "pg - employment in school", "pg - employment job placement",
  "pg - employment job training", "dc - wise ged", "dc - wise job placement",
  "dc - wise job training"))
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data."[Admin] raw_enrollment_report"}

\item{job_data}{dataframe: a dataframe containing job data."[Admin] raw_pos_report_detailed"}

\item{measure}{character vector: Name of measures in the table}

\item{target}{numeric vector: Organization target for each measure in the table}

\item{eto_programs}{character vector:  a vector of character containing the name of ETO programs to keep for GED analysis.}

\item{workforce_programs}{character vector:  a vector of character containing the name of all workforce programs.}
}
\value{
dataframe
}
\description{
This function creates a summary table of GED results.
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
enroll <- laycUtils::format_data(enroll)
job <- laycUtils::load_txt('./my_data_folder/job.txt')
job <- laycUtils::format_data(tp)

job_table(enroll_data = enroll, job_data = job)
}

