% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03_jrt_table.R
\name{jrt_table}
\alias{jrt_table}
\title{jrt_table()}
\usage{
jrt_table(enroll_data, tp_data, measure = c("enrolled in JRT",
  "assessed on soft skills", "gained soft-skills"), target = c(enrolled =
  0.75, assessed = 0.9, gained = 0.7), eto_programs = c("ss - job readiness",
  "pg - employment job training", "dc - wise job training"),
  workforce_programs = c("ss - ccorps projects", "ss - counseling",
  "ss - ged", "ss - job placement", "ss - job readiness",
  "pg - employment case management", "pg - employment ged",
  "pg - employment in school", "pg - employment job placement",
  "pg - employment job training", "dc - wise ged", "dc - wise job placement",
  "dc - wise job training"), for_bullet = "no")
}
\arguments{
\item{enroll_data}{dataframe: a dataframe containing enrollment data."[Admin] raw_enrollment_report"}

\item{tp_data}{dataframe: a dataframe containing touchpoint data."[Admin] raw_touchpoint_report_detailed"}

\item{eto_programs}{character vector:  a vector of character containing the name of ETO programs to keep for analysis.}
}
\value{
dataframe
}
\description{
This function creates a summary table of JRT results.
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
enroll <- laycUtils::format_data(enroll)
tp <- laycUtils::load_txt('./my_data_folder/touchpoints.txt')
tp <- laycUtils::format_data(tp)

jrt_table(enroll_data = enroll, tp_data = tp)
}

