% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_weights}
\alias{sum_weights}
\title{sum_weights()}
\usage{
sum_weights(tp_data, var = c("subject_id", "date", "answer_id", "question_id",
  "question_short", "answer_weight"), weight_var = "answer_weight",
  weight_id = "^q", group_var = c("subject_id", "answer_id"))
}
\arguments{
\item{tp_data}{dataframe: a dataframe containing touchpoint data."[Admin] raw_touchpoint_report_detailed"}

\item{var}{character vector: column names to keep for calculations}

\item{weight_id}{character: pattern that identifies the question to be included in the sum of weights. Can be  a regular expression.}
}
\value{
dataframe
}
\description{
This is a helper function that sums question weights from surveys to compute a total score. The data comes from the following ETO Results report: "[Admin] raw_touchpoint_report_detailed".
}
\examples{
tp <- laycUtils::load_txt('./my_data_folder/touchpoints.txt')
tp <- laycUtils::format_data(tp)
tp <- tp[tp$tp_name == "jrt pre-/post test", ]

sum_weights(tp_data = tp)
}

